+++ src/universal/withGraphQLApp.mjs
@@ -8,6 +8,23 @@ import Head from 'next/head'
 import React from 'react'
 
 /**
+ * Link `rel` types that make sense to forward from a GraphQL responses during
+ * SSR in the Next.js page response
+ * @see [HTML Living Standard link types](https://html.spec.whatwg.org/dev/links.html#linkTypes).
+ * @kind constant
+ * @name FORWARDABLE_LINK_REL
+ * @ignore
+ */
+const FORWARDABLE_LINK_REL = [
+  'dns-prefetch',
+  'preconnect',
+  'prefetch',
+  'preload',
+  'modulepreload',
+  'prerender'
+]
+
+/**
  * A higher-order React component to decorate a Next.js custom `App` component
  * in `pages/_app.js` for [`graphql-react`](https://npm.im/graphql-react),
  * enabling descendant GraphQL operations with server side rendering and client
@@ -15,8 +32,18 @@ import React from 'react'
  *
  * It also forwards HTTP
  * [`Link`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Link)
- * `rel="preload"` headers from GraphQL responses during SSR in the Next.js page
- * response. Link URLs are forwarded unmodified, so avoid sending relative URLs
+ * headers with the following `rel` parameters from GraphQL responses received
+ * when [`ssr`](https://github.com/jaydenseric/graphql-react#function-ssr) runs
+ * to the Next.js page response:
+ *
+ * - [`dns-prefetch`](https://html.spec.whatwg.org/dev/links.html#link-type-dns-prefetch)
+ * - [`preconnect`](https://html.spec.whatwg.org/dev/links.html#link-type-preconnect)
+ * - [`prefetch`](https://html.spec.whatwg.org/dev/links.html#link-type-prefetch)
+ * - [`preload`](https://html.spec.whatwg.org/dev/links.html#link-type-preload)
+ * - [`modulepreload`](https://html.spec.whatwg.org/dev/links.html#link-type-modulepreload)
+ * - [`prerender`](https://html.spec.whatwg.org/dev/links.html#link-type-prerender)
+ *
+ * Link URLs are forwarded unmodified, so avoid sending relative URLs
  * from a GraphQL server hosted on a different domain to the app.
  * @see [Next.js custom `App` docs](https://nextjs.org/docs#custom-app).
  * @see [React higher-order component docs](https://reactjs.org/docs/higher-order-components).
@@ -126,12 +153,11 @@ export const withGraphQLApp = App =>
 
                 graphqlLinkHeader.refs.forEach(graphqlLink => {
                   if (
-                    graphqlLink.rel === 'preload' &&
-                    // Identical preload link not already set.
+                    FORWARDABLE_LINK_REL.includes(graphqlLink.rel) &&
+                    // Similar link not already set.
                     !responseLinkHeader.refs.some(
-                      responseLink =>
-                        responseLink.rel === 'preload' &&
-                        responseLink.uri === graphqlLink.uri
+                      ({ uri, rel }) =>
+                        uri === graphqlLink.uri && rel === graphqlLink.rel
                     )
                   )
                     responseLinkHeader.set(graphqlLink)